requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# Image handling
Pillow==10.1.0

# Filtering
django-filter==23.3

# API Documentation
drf-yasg==1.21.7

# CORS headers
django-cors-headers==4.3.0

# Environment variables
python-decouple==3.8

# Timezone support
pytz==2023.3

# For Phase 2+
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# For Phase 3
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# For reporting
openpyxl==3.1.2
reportlab==4.0.7

# Security
django-ratelimit==4.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Production server
gunicorn==21.2.0

# Development
django-debug-toolbar==4.2.0

